refactor: 랭킹 배치 패키지와 네이밍 정리
배치 전용 랭킹 서비스를 batch 하위로 이동하고 리포지토리 및 스케줄러 관련 네이밍을 일관되게 정리한다.
