+++ src/main/java/com/LetMeDoWith/LetMeDoWith/domain/ranking/repository/RankingTopicRepository.java
@@ -4,7 +4,7 @@ import com.LetMeDoWith.LetMeDoWith.domain.ranking.model.RankingTopic;
 import java.util.Optional;
 
 public interface RankingTopicRepository {
-    Optional<RankingTopic> getActiveTopicByTitle(String title);
+    Optional<RankingTopic> getRankingTopic(String title);
 
     RankingTopic save(RankingTopic rankingTopic);
 }
